+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## Unreleased
 
+### Fixed
+- `autocomplete` Fixed  issues with items with identical labels and a custom template.
 
 ## [2.0.1] - 2019-04-04